+++ R/BITR_geocode.R
@@ -1,6 +1,6 @@
-#' Georeferenced data from the journal BioTropica (pulled from Web of Science)
+#' Georeferenced data from the journal Biotropica (pulled from Web of Science)
 #'
-#' A dataset containing 41 authors taken from the BioTropica journal. 
+#' A dataset containing 41 authors taken from the Biotropica journal. 
 #' This dataset represents the typical formatted output 
 #' from \code{authors_georef()}
 #' in the refsplitr package. It serves as a useful testing data set for
@@ -16,8 +16,8 @@
 #'   \item{lat}{numeric, latitude populated from authors_georef}
 #'   \item{lon}{numeric, longitude populated from authors_georef}
 #'   \item{groupID}{ID field for what name group the author 
-#'   is identied as from authors_clean()}
-#'   \item{author_order}{numeric, order of author from jounral article}
+#'   is identified as from authors_clean()}
+#'   \item{author_order}{numeric, order of author from journal article}
 #'   \item{address}{address of references pulled from 
 #'   the original raw WOS file}
 #'   \item{department}{department which is nested within university}