+++ R/references_read.R
@@ -2,11 +2,11 @@
 #'
 #' \code{references_read} This function reads Thomson Reuters Web of Knowledge
 #' and ISI format reference data files into an R-friendly data format. The resulting dataframe
-#' is the argument for the refplitr function `authors_clean()`.    
+#' is the argument for the refsplitr function `authors_clean()`.    
 #'
 #' @param data the location of the file or files to be imported. This can be either the absolute or 
 #' relative name of the file (for a single file) or folder (for multiple files stored in the same folder; 
-#' used in conjuction with `dir = TRUE``). If left blank it is assumed the location is the working directory.
+#' used in conjunction with `dir = TRUE``). If left blank it is assumed the location is the working directory.
 #' @param dir if FALSE it is assumed a single file is to be imported. 
 #' Set to TRUE if importing multiple files (the path to the folder in which files are stored is set with `data=``; 
 #' all files in the folder will be imported). Defaults to FALSE.